fix: remove sympy
