+++ pyproject.toml
@@ -30,7 +30,6 @@ dependencies = [
     'superqt >=0.3.1',
     'fonticon-materialdesignicons6',
     'qtpy',
-    'sympy'
 ]
 
 # extras
