+++ src/pymmcore_widgets/_hcs_widget/_main_hcs_widget.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import warnings
 from typing import List, Optional, Tuple, Union
